+++ helper-scripts/weather.sh
@@ -1,0 +1,78 @@
+#!/bin/bash
+#V0.0.8 - Beta
+
+# load in configuration variables
+. config-temp.conf
+#test variable run yes/no
+#convert variable to lowercase
+processweather1=$(echo $processweather | tr '[:upper:]' '[:lower:]')
+
+if [[ $processweather1 = yes ]]
+then
+
+#weather
+
+#audio
+randomNumber=$(shuf -i 1-7 -n 1 --repeat)
+randomNumber1=$(shuf -i 1-7 -n 1 --repeat)
+randomNumber2=$(shuf -i 1-7 -n 1 --repeat)
+audio=$(head -n $randomNumber $workdir/music.txt | tail -n 1)
+audio1=$(head -n $randomNumber1 $workdir/music.txt | tail -n 1)
+audio2=$(head -n $randomNumber2 $workdir/music.txt | tail -n 1)
+
+
+#Backgrounds
+
+#background randomisation
+if [[ $backgroundcolour == random ]]
+then
+awk 'BEGIN{srand()}{print rand(), $0}' $workdir/colours.txt | sort -n -k 1 | awk 'sub(/\S* /,"")'
+backgroundrandomNumber=$(shuf -i 1-140 -n 1 --repeat)
+backgroundrandomNumber1=$(shuf -i 1-140 -n 1 --repeat)
+backgroundrandomNumber2=$(shuf -i 1-140 -n 1 --repeat)
+background=$(head -n $backgroundrandomNumber $workdir/colours.txt | tail -n 1)
+background1=$(head -n $backgroundrandomNumber1 $workdir/colours.txt | tail -n 1)
+background2=$(head -n $backgroundrandomNumber2 $workdir/colours.txt | tail -n 1)
+echo $background
+echo $background1
+else
+background=$backgroundcolour
+background1=$backgroundcolour
+background2=$backgroundcolour
+fi
+
+#set weather celcius or farenheit using country code
+if [[ $country == US ]]
+then
+weathermeasurement=?u
+else
+weathermeasurement=?m
+fi
+
+
+
+#retrieve weather data
+
+#fix white spaces for curl
+
+stateurl=$(echo $state|sed -e 's/ /%20/g')
+cityurl=$(echo $city|sed -e 's/ /%20/g')
+
+
+
+curl wttr.in/${cityurl}.png$weathermeasurement --output $weatherdir/v1.png
+curl v2.wttr.in/${cityurl}.png$weathermeasurement --output $weatherdir/v2.png
+curl v3.wttr.in/${stateurl}.png$weathermeasurement --output $weatherdir/v3.png
+wait
+
+#make video
+
+ffmpeg -y -f lavfi -i color=$background:$videoresolution:d=$videolength -i $weatherdir/v1.png -stream_loop -1 -i $audio -shortest -filter_complex "[1]scale=iw*1:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy $output/weather-v1.mp4
+ffmpeg -y -f lavfi -i color=$background1:$videoresolution:d=$videolength -i $weatherdir/v2.png -stream_loop -1 -i $audio1 -shortest -filter_complex "[1]scale=iw*1:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy $output/weather-v2.mp4
+ffmpeg -y -f lavfi -i color=$background2:$videoresolution:d=$videolength -i $weatherdir/v3.png -stream_loop -1 -i $audio2 -shortest -filter_complex "[1]scale=iw*0.9:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy $output/weather-v3.mp4
+
+#end weather
+./news.sh
+else
+./news.sh
+fi